rename: 게임 사이드바 컴포넌트명 변경
