+++ src/components/game/GameSidebar.tsx
